refactor(showcase,roadmap,i18n): 提取共用選單組件並修正 review feedback
## Why is this necessary?

- BrewingCard 與 PracticeShowcaseCard 的擁有者選單完全複製貼上，觸發 SonarCloud 74.5% duplication Quality Gate 失敗
- onRestore 失敗時錯誤提示使用 archive_failed（封存失敗）而非 restore_failed（復原失敗），誤導使用者
- zh-TW.json 隱私權政策內容中「穩私權」為錯字，應為「隱私權」
- roadmap 許願 CTA 使用原生 <button> 而非專案 Button 組件，缺少統一的 focus/a11y 行為

## How does it address?

- 新增 PracticeActionMenu 共用組件，封裝擁有者/非擁有者選單的 UI、狀態管理、click-outside 與所有 action handler
- BrewingCard 與 PracticeShowcaseCard 改用 PracticeActionMenu，各移除 ~150 行重複程式碼
- onRestore catch 內改用 practiceT("restore_failed") 取代 practiceT("archive_failed")
- 修正 zh-TW.json privacy_policy_content 中「此穩私權政策」→「此隱私權政策」
- roadmap-board.tsx 的許願 CTA 從原生 <button> 改為 <Button> 組件
