+++ apps/product/src/components/showcase/PracticeShowcaseCard.tsx
@@ -1,52 +1,22 @@
 "use client";
 
-import {
-  type BatchReactionItem,
-  followTarget,
-  unfollowTarget,
-  useArchivePractice,
-  useComments,
-  useCopyPractice,
-  useCurrentUser,
-  useDeletePractice,
-  usePracticeById,
-} from "@daodao/api";
-import {
-  ChartColumnIncreasingSvg,
-  DefaultAvatarSvg,
-  DialogOutlineSvg,
-  FlagOutlineSvg,
-  TelescopeSvg,
-} from "@daodao/assets";
-import { useAuth } from "@daodao/auth";
+import { type BatchReactionItem, useComments, useCurrentUser, usePracticeById } from "@daodao/api";
+import { DefaultAvatarSvg, DialogOutlineSvg } from "@daodao/assets";
 import { useLocale, useTranslations } from "@daodao/i18n";
-import { Link, usePathname, useRouter } from "@daodao/i18n/navigation";
+import { Link, useRouter } from "@daodao/i18n/navigation";
 import { useSheetManager } from "@daodao/ui/components/animate-ui/components/radix/sheet";
 import { Avatar, AvatarFallback, AvatarImage } from "@daodao/ui/components/avatar";
 import { Badge } from "@daodao/ui/components/badge";
-import { Button } from "@daodao/ui/components/button";
-import { toast } from "@daodao/ui/components/sonner";
-import { cn } from "@daodao/ui/lib/utils";
-import { Archive, Copy, MoreHorizontal, Pencil, Trash2 } from "lucide-react";
-import { useEffect, useRef, useState } from "react";
 import { ReactionPickerButton } from "@/components/check-in/reactions";
 import {
   BrowseActivityContent,
   type IBrowseActivityFollower,
 } from "@/components/practice/shared/browse-activity-content";
-import { refreshOnboardingStatus } from "@/components/task-guide/onboarding-progress-context";
-import {
-  ArchivePracticeResult,
-  useArchivePracticeDialog,
-} from "@/hooks/use-archive-practice-dialog";
-import {
-  DeletePracticeResult,
-  useDeletePracticeDialog,
-} from "@/hooks/use-delete-practice-dialog";
 import type { ReactionTypeType } from "@/constants/reaction-type";
 import { getStatusConfig, TaskStatus } from "@/constants/task-status";
 import { useCardReactions } from "@/hooks/use-card-reactions";
 import { formatRelativeTime } from "@/utils/format-time";
+import { PracticeActionMenu } from "./PracticeActionMenu";
 import { formatShowcaseDate } from "./utils";
 
 interface PracticeShowcaseCardProps {
@@ -96,7 +66,6 @@ export function PracticeShowcaseCard({
   onReactionMutate,
 }: PracticeShowcaseCardProps) {
   const t = useTranslations("app_product");
-  const practiceT = useTranslations("practice");
   const commonT = useTranslations("common");
   const locale = useLocale();
   const startFmt = formatShowcaseDate(startDate);
@@ -104,65 +73,14 @@ export function PracticeShowcaseCard({
   const taskStatus = status === "active" ? TaskStatus.inProgress : TaskStatus.completed;
   const statusInfo = getStatusConfig(taskStatus);
 
-  const [menuOpen, setMenuOpen] = useState(false);
-  const [isFollowing, setIsFollowing] = useState(false);
-  const [isCopying, setIsCopying] = useState(false);
-  const menuRef = useRef<HTMLDivElement>(null);
-  const { isAuthenticated } = useAuth();
   const router = useRouter();
-  const pathname = usePathname();
   const { open: openSheet, close: closeSheet } = useSheetManager();
   const { data: practiceData } = usePracticeById(id);
   const { data: commentsData } = useComments({ targetType: "practice", targetId: id });
   const { data: currentUserData } = useCurrentUser();
   const isOwner = !!currentUserData?.data?.id && user?.id === currentUserData.data.id;
 
-  const { copyPractice } = useCopyPractice();
-  const { archivePractice, restorePractice } = useArchivePractice(id);
-  const { deletePractice } = useDeletePractice(id);
-  const { openArchiveDialog } = useArchivePracticeDialog();
-  const { openDeleteDialog } = useDeletePracticeDialog();
-
-  useEffect(() => {
-    if (!menuOpen) return;
-    const handleClickOutside = (e: MouseEvent | TouchEvent) => {
-      if (menuRef.current && !menuRef.current.contains(e.target as Node)) {
-        setMenuOpen(false);
-      }
-    };
-    document.addEventListener("mousedown", handleClickOutside);
-    document.addEventListener("touchstart", handleClickOutside);
-    return () => {
-      document.removeEventListener("mousedown", handleClickOutside);
-      document.removeEventListener("touchstart", handleClickOutside);
-    };
-  }, [menuOpen]);
-
-  const handleToggleFollow = async () => {
-    if (!isAuthenticated) {
-      const search = typeof window !== "undefined" ? window.location.search : "";
-      const redirectUrl = search ? `${pathname}${search}` : pathname;
-      router.push(`/auth/login?redirect=${encodeURIComponent(redirectUrl)}`);
-      return;
-    }
-    const wasFollowing = isFollowing;
-    setIsFollowing(!wasFollowing);
-    try {
-      if (wasFollowing) {
-        await unfollowTarget("practice", id);
-        toast.success(t("following_unfollowed"));
-      } else {
-        await followTarget({ targetType: "practice", targetId: id });
-        toast.success(t("following_followed_practice"));
-      }
-    } catch {
-      setIsFollowing(wasFollowing);
-      toast.error(t("operation_failed_retry"));
-    }
-  };
-
   const handleOpenBrowseActivity = () => {
-    setMenuOpen(false);
     const followers: IBrowseActivityFollower[] = reactionItems.map((item) => ({
       id: item.userId,
       name: item.name,
@@ -219,158 +137,11 @@ export function PracticeShowcaseCard({
         )}
 
         {/* More menu */}
-        {/* biome-ignore lint/a11y/useKeyWithClickEvents: stop card click */}
-        {/* biome-ignore lint/a11y/noStaticElementInteractions: stop card click */}
-        <div ref={menuRef} className="relative" onClick={(e) => e.stopPropagation()}>
-          <Button
-            type="button"
-            size="icon"
-            variant="ghost"
-            onClick={() => setMenuOpen((v) => !v)}
-            className={cn("h-8 w-8", menuOpen ? "bg-[#E4EAE9]" : "hover:bg-[#E4EAE9]")}
-          >
-            <MoreHorizontal className="size-4" />
-          </Button>
-
-          {menuOpen && isOwner && (
-            <div className="absolute right-0 top-full mt-1 bg-white rounded-2xl shadow-lg border border-[#E4EAE9] py-2 z-20 min-w-[140px]">
-              <Button
-                type="button"
-                variant="ghost"
-                onClick={() => {
-                  setMenuOpen(false);
-                  router.push(`/practices/${id}/edit`);
-                }}
-                className="w-full h-auto justify-start rounded-none gap-3 px-4 py-3 text-sm text-[#295E5C] hover:bg-[#F0F9F8] transition-colors cursor-pointer"
-              >
-                <Pencil className="size-[18px] shrink-0" />
-                <span>{practiceT("action_edit")}</span>
-              </Button>
-              <Button
-                type="button"
-                variant="ghost"
-                disabled={isCopying}
-                onClick={async () => {
-                  setMenuOpen(false);
-                  try {
-                    setIsCopying(true);
-                    const { id: newId } = await copyPractice(id);
-                    refreshOnboardingStatus();
-                    router.push(`/practices/copy-success?practiceId=${newId}`);
-                  } catch {
-                    toast.error(practiceT("copy_failed"));
-                  } finally {
-                    setIsCopying(false);
-                  }
-                }}
-                className="w-full h-auto justify-start rounded-none gap-3 px-4 py-3 text-sm text-[#295E5C] hover:bg-[#F0F9F8] transition-colors cursor-pointer"
-              >
-                <Copy className="size-[18px] shrink-0" />
-                <span>{practiceT("action_copy")}</span>
-              </Button>
-              <Button
-                type="button"
-                variant="ghost"
-                onClick={async () => {
-                  setMenuOpen(false);
-                  const result = await openArchiveDialog({
-                    onRestore: async () => {
-                      try {
-                        await restorePractice();
-                        toast.success(practiceT("restore_success"));
-                      } catch {
-                        toast.error(practiceT("archive_failed"));
-                      }
-                    },
-                  });
-                  if (result === ArchivePracticeResult.Archived) {
-                    try {
-                      await archivePractice();
-                      router.refresh();
-                    } catch {
-                      toast.error(practiceT("archive_failed"));
-                    }
-                  }
-                }}
-                className="w-full h-auto justify-start rounded-none gap-3 px-4 py-3 text-sm text-[#295E5C] hover:bg-[#F0F9F8] transition-colors cursor-pointer"
-              >
-                <Archive className="size-[18px] shrink-0" />
-                <span>{practiceT("action_archive")}</span>
-              </Button>
-              <Button
-                type="button"
-                variant="ghost"
-                onClick={handleOpenBrowseActivity}
-                className="w-full h-auto justify-start rounded-none gap-3 px-4 py-3 text-sm text-[#295E5C] hover:bg-[#F0F9F8] transition-colors cursor-pointer"
-              >
-                <ChartColumnIncreasingSvg className="size-[18px] shrink-0" />
-                <span>{practiceT("action_browse_activity")}</span>
-              </Button>
-              <Button
-                type="button"
-                variant="ghost"
-                onClick={async () => {
-                  setMenuOpen(false);
-                  const result = await openDeleteDialog();
-                  if (result === DeletePracticeResult.Deleted) {
-                    try {
-                      await deletePractice();
-                      router.refresh();
-                    } catch {
-                      toast.error(practiceT("delete_failed"));
-                    }
-                  }
-                }}
-                className="w-full h-auto justify-start rounded-none gap-3 px-4 py-3 text-sm text-red-500 hover:bg-red-50 transition-colors cursor-pointer"
-              >
-                <Trash2 className="size-[18px] shrink-0" />
-                <span>{practiceT("action_delete")}</span>
-              </Button>
-            </div>
-          )}
-          {menuOpen && !isOwner && (
-            <div className="absolute right-0 top-full mt-1 bg-white rounded-2xl shadow-lg border border-[#E4EAE9] py-2 z-20 min-w-[140px]">
-              <Button
-                type="button"
-                variant="ghost"
-                onClick={() => {
-                  setMenuOpen(false);
-                  window.open("https://tally.so/r/BzGQy4", "_blank");
-                }}
-                className="w-full h-auto justify-start rounded-none gap-3 px-4 py-3 text-sm text-[#295E5C] hover:bg-[#F0F9F8] transition-colors cursor-pointer"
-              >
-                <FlagOutlineSvg className="size-5 shrink-0" />
-                <span>{commonT("report")}</span>
-              </Button>
-              <Button
-                type="button"
-                variant="ghost"
-                onClick={() => {
-                  setMenuOpen(false);
-                  void handleToggleFollow();
-                }}
-                className={cn(
-                  "w-full h-auto justify-start rounded-none gap-3 px-4 py-3 text-sm transition-colors cursor-pointer",
-                  isFollowing
-                    ? "text-logo-cyan hover:bg-[#E8FAF9]"
-                    : "text-[#295E5C] hover:bg-[#F0F9F8]"
-                )}
-              >
-                <TelescopeSvg className="size-5 shrink-0" />
-                <span>{isFollowing ? t("following_unfollow") : t("following_follow")}</span>
-              </Button>
-              <Button
-                type="button"
-                variant="ghost"
-                onClick={handleOpenBrowseActivity}
-                className="w-full h-auto justify-start rounded-none gap-3 px-4 py-3 text-sm text-[#295E5C] hover:bg-[#F0F9F8] transition-colors cursor-pointer"
-              >
-                <ChartColumnIncreasingSvg className="size-5 shrink-0" />
-                <span>{t("showcase_browse_activity")}</span>
-              </Button>
-            </div>
-          )}
-        </div>
+        <PracticeActionMenu
+          practiceId={id}
+          isOwner={isOwner}
+          onBrowseActivity={handleOpenBrowseActivity}
+        />
       </div>
 
       {/* Title */}
