+++ packages/i18n/src/locales/zh-TW.json
@@ -478,7 +478,7 @@
   },
   "terms": {
     "privacy_policy_title": "隱私權政策",
-    "privacy_policy_content": "# 島島阿學隱私權政策\n\n島島阿學致力於重塑學習的本質，我們相信學習應該是自主的、有意義的，並與真實世界緊密連結。在這個瞬息萬變的時代，我們打造一個讓每個人都能持續成長的學習生態圈。我們的註冊使用者（「會員」）分享他們的學習經驗、學習資源並與其他會員進行學習交流活動，展現個人技能、經歷與成長，發佈和查看相關內容，並尋找可能的共同成長及合作機會。非會員（「訪客」）可以查看我們某些服務的內容和資料。\n\n此穩私權政策存在於您及本網站管理機關島島阿學（https://daodao.so ，以下簡稱「本網站」）間。 我們的隱私權政策適用於我們服務的任何註冊使用者或訪客。  \n請閱讀以下條款及條件並確認，當您上傳內容至本網站時，即表示您接受本協議內容。\n\n## 個人資料\n\n網站會員登入後，島島阿學學取得會員在第三方服務例如 Google的電子郵件地址與應用程式使用者 ID，用以分辨不同帳號以及聯繫會員。顯示名稱與頭像等公開顯示資訊，作為 島島阿學學習社群網站的預設頭像與顯示名稱。  \n網站會員登入後的新增之資源、主題實踐、回應內容、使用本平台之頻率等之公開訊息，將可能為島島阿學團隊用於數據分析，當進行量化或去識別化等過程與原身分識別勾脫後，會留存於開放資料隱去名稱之研究或統計。  \n本網站會嚴格保護會員之隱私及個人資料，除充分量化或去識別化後得不再視為個人資料者外，將不基於任何目的使其外流至與本服務無關之第三方。\n\n網站會員可在個人頁面逕行變更自己在島島阿學網站的預設頭像、顯示名稱及其他社群平台資訊，使其不再與 第三方服務上的公開資訊相同。另外，會員亦可寄信至 本網站連絡信箱（contact@daodao.so），請求刪除或更換取自第三方服務的公開顯示資訊或電子郵件地址。 本網站會在確認收到請求後的最遲三十天內，比對驗證寄件者電子郵件地址無誤，足證申請者為原個人資料當事人之後進行刪除或更換處理。\n\n當您註冊帳戶時，我們可能會要求您提供相關資訊，例如您的教育背景、工作經驗、技能、照片、城市或地區。個人資料資訊可幫助我們提供給您更多服務。您可以選擇是否在您的個人資料中包含敏感資訊並公開該敏感資訊。請勿在您的個人資料中發布或新增您不希望公開的個人資料。\n\n## 第三方服務整合\n\n為提供更完整的功能與服務，本服務可能與第三方平台或服務進行整合，包括但不限於Google服務（如日曆整合，用於學習時間安排）和其他社交媒體或認證服務。當您選擇啟用相關整合功能時，可能會授權我們存取、蒐集及使用您於第三方服務中的特定資訊（例如帳號資料、聯絡人、日曆事件等），以實現登入驗證、資料同步或其他您所需的服務功能。我們僅會在獲得您的明確同意後，根據本政策所述目的使用這些資訊，並嚴格遵循各第三方服務的使用條款與限制。我們僅暫時快取必要資訊以提供服務，不會永久儲存您的內容，僅用於改善您的學習體驗。\n\n### Google 服務特別說明\n\n當您連接Google日曆時，我們會透過安全的API存取您的日曆事件資訊，僅用於協助您安排學習時間和提供個人化學習建議。我們不會將您的Google資料分享給其他第三方，且您可隨時在設定中撤銷此授權。我們的Google資料使用完全符合Google API服務使用者資料政策的有限使用要求。\n\n## 來自第三方網站的嵌入內容\n\n這個網站上的文章可能會嵌入視訊、圖片、文章等內容，而來自第三方網站的嵌入內容，其隱私權處理方式與使用者造訪這些網站時的規定完全相同。無論使用者是否有這些第三方網站的帳號或是否登入網站，他們都會以各種方式收集與使用者相關的資料，如 Cookie、嵌入第三方追蹤程式碼、監視使用者與嵌入內容的互動等。\n\n來自第三方來源的資訊\n\n我們也可能從其他來源獲取您的資訊。例如，我們可能從第三方（包括但不限於郵件列表提供商和公開來源）收集您的資訊。當您在服務上與我們的會員和線上社群互動時，我們可能會從他們那裡收集您選擇揭露或提供的任何內容或資訊。我們也可能從其他來源獲取您的資訊，包括但不限於社群媒體平台、行銷合作夥伴和其他第三方。例如，如果您將您的社群媒體帳戶連結到您的帳戶，或者透過社群媒體網站建立或登入您的帳戶，您的社群媒體帳戶的某些資訊將會揭露給我們，其中可能包括您個人資料或您朋友個人資料中的資訊。\n\n## 使用者資料分析\n\n島島阿學平台使用人工智慧技術(AI)來優化用戶使用體驗。為了提供這些功能，我們會處理用戶資料，包括學習活動紀錄、資源互動情形及內容貢獻。我們將使用匿名化及去識別化的資料來訓練和改進我們的 AI 系統。在任何資料用於 AI 訓練之前，個人識別資訊都會被移除。使用者產生的內容，如學習資源和討論，可能會被我們的 AI 系統分析，以改善內容推薦和用戶互動體驗。\n\n* AI 生成內容  \n  島島阿學平台上的 AI 生成內容可能包括：  \n- **平台生成內容**：由島島阿學平台的AI系統自動生成的內容，如學習推薦或覆盤回饋等。這些內容將由平台明確標示為 AI 生成，以確保透明度。  \n- **使用者上傳內容**：使用者上傳的任何內容，無論是否使用AI工具輔助創作，均視為使用者的原創貢獻。本平台目前無法且不會主動判斷或標示使用者上傳內容的創作方式。\n\n* AI決策過程  \n  我們的平台可能使用 AI 來提供個人化學習路徑和內容推薦。這些推薦是基於您的學習行為、興趣和目標。\n\n* AI安全與透明度  \n  島島阿學致力於負責任地使用AI技術。我們實施多項措施確保我們的 AI 系統不含偏見並避免產生有害內容。我們定期評估和測試 AI 模型，並根據使用者反饋進行調整。我們承諾透明地告知您 AI 功能的使用方式及其對您體驗的影響。\n\n## 這個網站的個人資料保留期限\n\n當使用者在這個網站發佈留言後，該則留言及其中繼資料將會無限期保留。這樣系統便可以自動辨識及核准任何後續留言，而不須將其保留在待審核的佇列中。針對在這個網站上註冊的使用者，這個網站還會儲存他們在使用者 \\[個人資料\\] 頁面中提供的個人資訊。全部使用者都可以隨時查看、編輯或刪除自己的個人資訊 (無法變更的使用者名稱除外)。請注意，網站管理員也可以查看及編輯這些個人資訊。\n\n## 使用者對個人資料擁有哪些權利\n\n如果使用者在這個網站擁有帳戶或曾發佈留言，便可以要求下載使用者在這個網站上的個人資料的資料匯出檔，這個檔案包含使用者提供給這個網站的全部個人資料。使用者也可以要求清除曾提供給這個網站的全部個人資料，但這項要求不包含站方為了管理、法律或安全目的而必須保留的相關資料。",
+    "privacy_policy_content": "# 島島阿學隱私權政策\n\n島島阿學致力於重塑學習的本質，我們相信學習應該是自主的、有意義的，並與真實世界緊密連結。在這個瞬息萬變的時代，我們打造一個讓每個人都能持續成長的學習生態圈。我們的註冊使用者（「會員」）分享他們的學習經驗、學習資源並與其他會員進行學習交流活動，展現個人技能、經歷與成長，發佈和查看相關內容，並尋找可能的共同成長及合作機會。非會員（「訪客」）可以查看我們某些服務的內容和資料。\n\n此隱私權政策存在於您及本網站管理機關島島阿學（https://daodao.so ，以下簡稱「本網站」）間。 我們的隱私權政策適用於我們服務的任何註冊使用者或訪客。  \n請閱讀以下條款及條件並確認，當您上傳內容至本網站時，即表示您接受本協議內容。\n\n## 個人資料\n\n網站會員登入後，島島阿學學取得會員在第三方服務例如 Google的電子郵件地址與應用程式使用者 ID，用以分辨不同帳號以及聯繫會員。顯示名稱與頭像等公開顯示資訊，作為 島島阿學學習社群網站的預設頭像與顯示名稱。  \n網站會員登入後的新增之資源、主題實踐、回應內容、使用本平台之頻率等之公開訊息，將可能為島島阿學團隊用於數據分析，當進行量化或去識別化等過程與原身分識別勾脫後，會留存於開放資料隱去名稱之研究或統計。  \n本網站會嚴格保護會員之隱私及個人資料，除充分量化或去識別化後得不再視為個人資料者外，將不基於任何目的使其外流至與本服務無關之第三方。\n\n網站會員可在個人頁面逕行變更自己在島島阿學網站的預設頭像、顯示名稱及其他社群平台資訊，使其不再與 第三方服務上的公開資訊相同。另外，會員亦可寄信至 本網站連絡信箱（contact@daodao.so），請求刪除或更換取自第三方服務的公開顯示資訊或電子郵件地址。 本網站會在確認收到請求後的最遲三十天內，比對驗證寄件者電子郵件地址無誤，足證申請者為原個人資料當事人之後進行刪除或更換處理。\n\n當您註冊帳戶時，我們可能會要求您提供相關資訊，例如您的教育背景、工作經驗、技能、照片、城市或地區。個人資料資訊可幫助我們提供給您更多服務。您可以選擇是否在您的個人資料中包含敏感資訊並公開該敏感資訊。請勿在您的個人資料中發布或新增您不希望公開的個人資料。\n\n## 第三方服務整合\n\n為提供更完整的功能與服務，本服務可能與第三方平台或服務進行整合，包括但不限於Google服務（如日曆整合，用於學習時間安排）和其他社交媒體或認證服務。當您選擇啟用相關整合功能時，可能會授權我們存取、蒐集及使用您於第三方服務中的特定資訊（例如帳號資料、聯絡人、日曆事件等），以實現登入驗證、資料同步或其他您所需的服務功能。我們僅會在獲得您的明確同意後，根據本政策所述目的使用這些資訊，並嚴格遵循各第三方服務的使用條款與限制。我們僅暫時快取必要資訊以提供服務，不會永久儲存您的內容，僅用於改善您的學習體驗。\n\n### Google 服務特別說明\n\n當您連接Google日曆時，我們會透過安全的API存取您的日曆事件資訊，僅用於協助您安排學習時間和提供個人化學習建議。我們不會將您的Google資料分享給其他第三方，且您可隨時在設定中撤銷此授權。我們的Google資料使用完全符合Google API服務使用者資料政策的有限使用要求。\n\n## 來自第三方網站的嵌入內容\n\n這個網站上的文章可能會嵌入視訊、圖片、文章等內容，而來自第三方網站的嵌入內容，其隱私權處理方式與使用者造訪這些網站時的規定完全相同。無論使用者是否有這些第三方網站的帳號或是否登入網站，他們都會以各種方式收集與使用者相關的資料，如 Cookie、嵌入第三方追蹤程式碼、監視使用者與嵌入內容的互動等。\n\n來自第三方來源的資訊\n\n我們也可能從其他來源獲取您的資訊。例如，我們可能從第三方（包括但不限於郵件列表提供商和公開來源）收集您的資訊。當您在服務上與我們的會員和線上社群互動時，我們可能會從他們那裡收集您選擇揭露或提供的任何內容或資訊。我們也可能從其他來源獲取您的資訊，包括但不限於社群媒體平台、行銷合作夥伴和其他第三方。例如，如果您將您的社群媒體帳戶連結到您的帳戶，或者透過社群媒體網站建立或登入您的帳戶，您的社群媒體帳戶的某些資訊將會揭露給我們，其中可能包括您個人資料或您朋友個人資料中的資訊。\n\n## 使用者資料分析\n\n島島阿學平台使用人工智慧技術(AI)來優化用戶使用體驗。為了提供這些功能，我們會處理用戶資料，包括學習活動紀錄、資源互動情形及內容貢獻。我們將使用匿名化及去識別化的資料來訓練和改進我們的 AI 系統。在任何資料用於 AI 訓練之前，個人識別資訊都會被移除。使用者產生的內容，如學習資源和討論，可能會被我們的 AI 系統分析，以改善內容推薦和用戶互動體驗。\n\n* AI 生成內容  \n  島島阿學平台上的 AI 生成內容可能包括：  \n- **平台生成內容**：由島島阿學平台的AI系統自動生成的內容，如學習推薦或覆盤回饋等。這些內容將由平台明確標示為 AI 生成，以確保透明度。  \n- **使用者上傳內容**：使用者上傳的任何內容，無論是否使用AI工具輔助創作，均視為使用者的原創貢獻。本平台目前無法且不會主動判斷或標示使用者上傳內容的創作方式。\n\n* AI決策過程  \n  我們的平台可能使用 AI 來提供個人化學習路徑和內容推薦。這些推薦是基於您的學習行為、興趣和目標。\n\n* AI安全與透明度  \n  島島阿學致力於負責任地使用AI技術。我們實施多項措施確保我們的 AI 系統不含偏見並避免產生有害內容。我們定期評估和測試 AI 模型，並根據使用者反饋進行調整。我們承諾透明地告知您 AI 功能的使用方式及其對您體驗的影響。\n\n## 這個網站的個人資料保留期限\n\n當使用者在這個網站發佈留言後，該則留言及其中繼資料將會無限期保留。這樣系統便可以自動辨識及核准任何後續留言，而不須將其保留在待審核的佇列中。針對在這個網站上註冊的使用者，這個網站還會儲存他們在使用者 \\[個人資料\\] 頁面中提供的個人資訊。全部使用者都可以隨時查看、編輯或刪除自己的個人資訊 (無法變更的使用者名稱除外)。請注意，網站管理員也可以查看及編輯這些個人資訊。\n\n## 使用者對個人資料擁有哪些權利\n\n如果使用者在這個網站擁有帳戶或曾發佈留言，便可以要求下載使用者在這個網站上的個人資料的資料匯出檔，這個檔案包含使用者提供給這個網站的全部個人資料。使用者也可以要求清除曾提供給這個網站的全部個人資料，但這項要求不包含站方為了管理、法律或安全目的而必須保留的相關資料。",
     "ipr_title": "智慧財產權",
     "ipr_content": "# 島島阿學智慧財產權\n\n島島阿學（https://daodao.so ，以下簡稱「本網站」）尊重他人的智慧財產權，希望提供一個不包含侵犯這些權利的內容平台。我們的使用者協議要求會員發布的資訊準確、合法且不侵犯第三方的權利，這個要求包含本網站所舉辦的任何型態活動。\n\n請注意，無論我們是否禁止存取或刪除內容，本網站都可能會善意地嘗試將書面通知（包括投訴人的聯絡資訊）轉發給發布該內容的會員和/或採取其他合理措施來通知該會員。表示本網站 已收到涉嫌侵犯智慧財產權或其他內容違規的通知。我們的政策也是，在適當的情況下，我們可以自行決定停用和/或終止侵犯或反覆侵犯他人權利或以其他方式發布非法內容的會員或團體（視情況而定）的帳戶。\n\n請注意，您提交的任何通知或反通知必須真實，且必須遵守偽證罪的處罰規定。虛假通知或反通知可能會導致個人責任。因此，您可能需要在提交通知或反通知之前尋求法律顧問的建議。\n\n## 著作權\n\n網站協作者查核後所撰寫的回應、補充或評價等資訊，若有產生著作權利保護之可能，撰寫之編輯同意採公眾領域宣告（CC0-1.0）無償提交至 本網站所維運之電腦或相關設備進行存放。  \n網站協作者若使用第三方之資料來源進行佐證，其引用之資料來源之著作仍屬於原創作之第三方所有，此時協作者同意為查證目的，在合理範圍內引用第三方資料，並於提交時提供原創作出處。  \n網站協作者及聊天機器人使用者採 CC0-1.0 提交資訊後，本網站將存放於電腦或其相關設備的該等資訊進行編輯性整理後，得採 CC BY-SA 4.0 或其他授權模式，將具編輯性保護之資料集發布於 島島阿學學習社群 網站、島島阿學學習社群 聊天機器人或 島島阿學學習社群 所提供資料存檔等處。  \n若資料利用者有違反前項授權模式的情事，將由 本網站主張權利。  \n本網站 將就社會公益及開放性進行綜合評估，來選擇發布資料時使用之授權模式，亦歡迎您透過 本網站 的聯繫方式，隨時將授權釋出政策有關的寶貴意見提供給我們。\n\n## 貢獻內容之合法及適當狀態\n\n您的資料貢獻不應侵害他人之智慧財產權利。如您貢獻內容，代表就您所知您表示，您有權授權管理者及本網站的使用者，依管理者指定、揀選的授權條款，來使用並散布這些內容。請注意管理者並不必然需要將您貢獻的內容包括於本網站上，並且得以在任何時候，將您的貢獻於本網站移除。對您貢獻的內容，管理者並無負擔保管責任。  \n透過AI功能輔助創作的內容，使用者仍需對最終發布的內容負責。\n\n## 授與權利\n\n您在此授與管理者全球性、免授權金、非專屬、永久、不可撤回之著作權及其他司法管轄區域可能定義之鄰接權及資料庫權之權利，讓其能夠不受前述權利限制使用內容之任何內含物，無論是就原始載體，或採其他型態進行利用。所授與的權利明示包括商業性使用，並且不排除任何領域內的利用。此一授權包括並不限於，將作品進行後續再授權利用，並允許多階的被授權人皆得採再授權方式再作利用。就現行法及著作權契約可容許的最大範圍內，您對管理者或經其授權利用內容之人，亦拋棄或不主張任何著作人格權。除本處訂明之範圍，您所貢獻內容之所有權利、地位，及利益，仍保留在您。\n\n## 免責聲明\n\n在現行法容許的範圍內，您是以現狀及現有之基礎提供本內容，但應保證該內容無侵權之虞。除依法不得事先排除或限縮之責任外，您或管理者皆不應依此協議，為其他歸責理論導致任何損害賠償。\n\n## 其他事項\n\n您所提送之表單及因之所需之註冊、登入程序裡，產生或由您填註的個人資訊，您在此同意管理者後續蒐集、處理及利用該個人資料。本「條款」依據中華民國法律。當本協議任一條款被視為無效時不影響其他部份之有效性。\n\n## 輔助條款\n\n除採前述貢獻提供內容外，您亦可透過表格、附錄，補充「描述」欄位說明的方式，表達您指定特一款或多款公眾授權條款，或宣告模式為輔助授權條款之意向。此時該等貢獻內容，除依本條款第2條規定之方式授權本網站外，同時亦依指定條款雙重／多重授權予本網站。您所指定的輔助條款，並不拘束管理者，但在處理方式容許的前提下，管理者得尊重您的指定，除依管理者指定、揀選的授權條款外，同時一併採用輔助條款來散布該等貢獻內容。倘您於輔助條款裡列入「CC0 公眾領域貢獻宣告」，該宣告模式為著作權及相關權利的拋棄聲明，解釋上本網站得不受任何著作權利與相關權利之限制，自由使用該等貢獻內容，並亦得再採 CC0 宣告模式散布該等貢獻內容。\n\n## 聯絡方式\n\n如果您對於本使用者條款、著作權、資料庫處理爭議、服務內容或隱私等有任何疑慮，請寄信至本網站之聯絡信箱 contact@daodao.so，由本網站在固定會議中或透過其他管道為您處理。紛爭解決和管轄法院。依本使用者條款所產生之爭議，雙方合意以臺灣臺北地方法院為第一審管轄法院。",
     "service_title": "使用者條款",
